requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
ffmpeg-python==0.2.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.28
langchain-text-splitters==1.1.1
langsmith==0.7.30
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.4
orjson==3.11.8
packaging==26.0
pgvector==0.4.2
pillow==12.2.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
shellingham==1.5.4
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
voyageai==0.3.7
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
